+++ projects/swimlane/ngx-charts/src/lib/common/axes/ticks.helper.ts
@@ -14,13 +14,21 @@ export function reduceTicks(ticks: any[], maxTicks: number): any[] {
 }
 
 export function getTickLines(label: string, maxLength: number, maxLines: number): string[] {
-  let totalLines = (label || '')
-    .toString()
-    .split(/\s+/)
-    .reduce((lines: string[], line: string) => {
+  const labelString = (label || '').toString();
+  let totalLines = [];
+
+  if (/\s/.test(labelString)) {
+    totalLines = labelString.split(/\s+/).reduce((lines: string[], line: string) => {
       const last = (lines.pop() || '') + ' ';
       return last.length + line.length > maxLength ? [...lines, last.trim(), line.trim()] : [...lines, last + line];
     }, []);
+  } else {
+    let startIndex = 0;
+    while (startIndex < labelString.length) {
+      totalLines.push(labelString.substring(startIndex, startIndex + maxLength));
+      startIndex += maxLength;
+    }
+  }
 
   if (totalLines.length > maxLines) {
     totalLines = totalLines.splice(0, maxLines);